update Controller & models

+++ src/controllers/bulletinController.js
@@ -4,21 +4,52 @@ import Bulletin from '../models/bulletin';
 import ImageArr from '../models/image_arr';
 import Hot from '../models/hot';
 import Like from '../models/like';
+import fs from 'fs';
 
 //---App---
 //Bulletin Inquiry
 export const bulletinInquiry = async(req, res) => {
+    const { page, limit } = req.query;
     try{
         const data = await Bulletin.findAll({
+            offset: (Number(page) - 1) * Number(limit),
+            limit: Number(limit),
             order: [['bulletin_id', 'DESC']],
         });
         
         return res.status(200).json(data);
     }catch (err) {
         console.error(err);
-        res.status(404);
+        return res.status(404);
     }
 };
+
+//Bulletin Detail
+export const bulletinDetail = async(req, res, next) => {
+    try{
+        const { bulletin_id } = req.query;
+        const data = await Bulletin.findOne({
+            where: {
+            bulletin_id,
+        },
+    });
+    return res.status(200).json(data);
+  } catch (err) {
+        console.error(err);
+        return res.status(404);
+    }
+}
+
+//Bulletin Count
+export const bulletinCount = async (req, res, next) => {
+    try {
+      const count = await Bulletin.count();
+      return res.status(200).json(count);
+    } catch (err) {
+        console.error(err);
+        return next();
+    }
+  };
 
 //Bulletin Image Inquiry
 export const bulletinImgInquiry = async(req, res) => {
@@ -38,20 +69,19 @@ export const bulletinImgInquiry = async(req, res) => {
 
 //Bulletin Create
 export const bulletinCreate = async(req, res, next) => {
-    const { title, content, images } = req.body;
-    try{
-        const data = await Bulletin.create({
+    const { title, content } = req.body;
+    try{ //게시글 생성
+        const bulletin = await Bulletin.create({
             title,
             content,
             create_date: Date.now(),
-            std_id: req.user.id,
+            std_id: req.user.std_id,
         });
-        if(images){
-            images.forEach(async (image) => {
-                await ImageArr.create({
-                    path: image.localUri,
-                    bulletin_id: data.dataValues.bulletin_id,
-                });
+        //이미지 배열에 경로 넣기
+        for(let file of req.files){
+            await ImageArr.create({
+                path: process.env.MY_DOMAIN + "/images/" + file.filename,
+                bulletin_id: bulletin.bulletin_id,
             });
         }
 
@@ -64,7 +94,7 @@ export const bulletinCreate = async(req, res, next) => {
 
 //Bulletin Update
 export const bulletinUpdate = async(req, res, next) => {
-    const { title, content, id, images } = req.body;
+    const { title, content, bulletin_id, should_delete_img } = req.body;
     try{
         await Bulletin.update(
             {
@@ -74,28 +104,42 @@ export const bulletinUpdate = async(req, res, next) => {
             },  
             {
                 where: {
-                    bulletin_id: id,
+                    bulletin_id,
                 },
             }
         );
-        await ImageArr.destroy({
-            where: {
-                bulletin_id: id,
-            },
-        });
-        if(images){
-            images.forEach(async (image) => {
-                await ImageArr.create(
-                    {
-                        path: image.localUri,
-                        bulletin_id: id,
-                    },
-                    {
-                        where: {
-                            bulletin_id: id,
-                        },
-                    }
-                );
+        //만약 삭제해야 될 이미지가 있으면 삭제
+        if (should_delete_img) {
+            if (typeof should_delete_img != "object") {
+              should_delete_img = [should_delete_img];
+            }
+            const imgArr = await ImageArr.findAll({
+              where: {
+                image_id: {
+                  [Op.in]: should_delete_img,
+                },
+              },
+            });
+      
+            try {
+              imgArr.forEach((el) => {
+                fs.unlinkSync("src/public/images/" + el.path.split("/")[4]);
+              });
+            } catch (err) {}
+      
+            await ImageArr.destroy({
+              where: {
+                image_id: {
+                  [Op.in]: should_delete_img,
+                },
+              },
+            });
+          }
+        //추가할 이미지가 있으면 추가
+        for (let file of req.files) {
+            await ImageArr.create({
+              path: process.env.MY_DOMAIN + "/images/" + file.filename,
+              bulletin_id,
             });
         }
 
@@ -107,14 +151,16 @@ export const bulletinUpdate = async(req, res, next) => {
 
 //Bulletin Search
 export const bulletinSearch = async(req, res, next) => {
-    const { title } = req.body;
+    const { keyword, page, limit } = req.query;
     try{
         const data = await Bulletin.findAll({
             where: {
                 title: {
-                    [Op.like]: '%' + title + '%', 
+                    [Op.like]: '%' + keyword + '%', 
                 },
             },
+            offset: (Number(page) - 1) * Number(limit),
+            limit: Number(limit),
             order: [['bulletin_id', 'DESC']],
         });
         
@@ -124,34 +170,49 @@ export const bulletinSearch = async(req, res, next) => {
     }
 };
 
-//hot Btn Click
+//Bulletin Search Count
+export const bulletinSearchCount = async (req, res, next) => {
+    const { keyword } = req.query;
+    try {
+      const count = await Bulletin.count({
+        where: {
+          title: {
+            [Op.like]: "%" + keyword + "%",
+          },
+        },
+      });
+      return res.status(200).json(count);
+    } catch (err) {
+        console.error(err);
+        return next();
+    }
+  };
+
+//hot Btn 
 export const bulletinClickHot = async(req, res) => {
-    const { id } = req.body;
-    try{
+    const { bulletin_id } = req.body;
+    try {
         const data = await Like.findOne({
-            where: { bulletin_id: id, std_id: req.user.id },
-        });
-        if(!data) {
-            await Like.create({ bulletin_id: id, std_id: req.user.id });
-            await Bulletin.increment(
-                { hot: 1 },
-                { where: { bulletin_id: id }}
-            );
+        where: { bulletin_id, std_id: req.user.std_id },
+    });
+        if (!data) {
+            await Like.create({ bulletin_id, std_id: req.user.std_id });
+            await Bulletin.increment({ hot: 1 }, { where: { bulletin_id } });
             const hotNum = await Bulletin.findOne({
-                attributes: ['hot'],
+                attributes: ["hot"],
                 where: {
-                    bulletin_id: id,
+                    bulletin_id,
                 },
             });
-            if(hotNum.dataValues.hot === 20) {
-                await Hot.create({
-                    bulletin_id: id,
-                });
-            }
+        if (hotNum.dataValues.hot === 20) {
+            await Hot.create({
+                bulletin_id,
+            });
+        }
 
             return res.status(200).json(hotNum);
         }else {
-            return res.status(200).send('한번 이상 클릭 하셨습니다.');
+            return res.status(400).send('한번 이상 클릭 하셨습니다.');
         }
     }catch (err) {
         return new Error(err);
@@ -163,7 +224,7 @@ export const bulletinInquiryView = async(req, res) => {
     try{
         await Bulletin.increment(
             { views: 1 },
-            { where: { bulletin_id: req.body.id }}
+            { where: { bulletin_id: req.body.bulletin_id }}
         );
         return res.status(200).send('Watch Success');
     }catch (err) {
@@ -173,22 +234,42 @@ export const bulletinInquiryView = async(req, res) => {
 };
 
 //Bulletin Delete
-export const bulletinDelete = async(req, res, next) => {
+export const bulletinDelete = async (req, res, next) => {
     const { bulletin_id } = req.body;
-    try{
-        await Bulletin.destroy({
-            where: {
-                bulletin_id,
-            },
+  
+    try {
+      const imgArr = await ImageArr.findAll({
+        where: {
+          bulletin_id,
+        },
+      });
+  
+      try {
+        imgArr.forEach((el) => {
+          console.log(el);
+          fs.unlinkSync("src/public/images/" + el.path.split("/")[4]);
         });
-        await Comment.destroy({
-            where: {
-                bulletin_id,
-            },
-        });
-        return res.status(200).send('Success');
-    }catch (err) {
-        console.error(err);
-        next(err);
+      } catch (err) {}
+  
+      await ImageArr.destroy({
+        where: {
+          bulletin_id,
+        },
+      });
+      await Bulletin.destroy({
+        where: {
+          bulletin_id,
+        },
+      });
+  
+      await Comment.destroy({
+        where: {
+          bulletin_id,
+        },
+      });
+      return res.status(200).send("Success");
+    } catch (err) {
+      console.error(err);
+      next(err);
     }
 };
